% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcBmsy.R
\name{calcBmsy}
\alias{calcBmsy}
\title{Calculate Bmsy by averaging over specified time intervals}
\usage{
calcBmsy(mmb, years = c(1980:1984, 1990:1997))
}
\arguments{
\item{mmb}{- time series of mmb at mating to average}

\item{years}{- years to incude in average}
}
\value{
Bmsy (in 1000's t)
}
\description{
Function to calculate Bmsy by averaging over specified time intervals
}
\details{

}

