% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcCIs.R
\name{calcCIs}
\alias{calcCIs}
\title{Compute confidence intervals}
\usage{
calcCIs(vals, cvs, pdfType = "lognormal", ci = 0.95)
}
\arguments{
\item{vals}{- vector of values}

\item{cvs}{- vector of cvs}

\item{pdfType}{- probability distribution for error bars}

\item{ci}{- confidence interval for error bar plots}
}
\description{
Function to compute confidence intervals.
}
\details{

}

