% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFofl.R
\name{calcFofl}
\alias{calcFofl}
\title{Calculate \eqn{F_{OFL}} using the Tier 4 harvest control rule}
\usage{
calcFofl(mmb, Bmsy, maxFofl, alpha, beta)
}
\arguments{
\item{mmb}{\itemize{
\item MMB at mating (from calcMMBmating())
}}

\item{Bmsy}{\itemize{
\item value for \eqn{B_{MSY}} (from clcBmsy())
}}

\item{maxFofl}{\itemize{
\item max allowed \eqn{F_{OFL} (= \gamma M)}
}}

\item{alpha}{\itemize{
\item value for the Tier 4 \eqn{\alpha} constant (the x-intercept of the sloping control rule)
}}

\item{beta}{\itemize{
\item value for the Tier 4 \eqn{\beta} constant (the threshold for \eqn{MMB/B_{MSY}} to allow directed fishing)
}}
}
\value{
the value of \eqn{F_{OFL}} given the mmb/Bmsy ratio, using the Tier 4 control rule.
}
\description{
Functio to calculate \eqn{F_{OFL}} using the Tier 4 harvest control rule.
}
\details{
none.
}
