% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcMMBMating.R
\name{calcMMBMating}
\alias{calcMMBMating}
\title{Calculate projected MMMB at mating.}
\usage{
calcMMBMating(avgSrvData, fshData, type = "averaged", M = 0.18,
  t.sf = 3/12, t.fm = 4/12, hm.pot = 0.5, hm.trl = 0.8,
  pct.male = 0.5)
}
\arguments{
\item{avgSrvData}{- dataframe with MMB at time of survey}

\item{fshData}{- dataframe with fishery data}

\item{type}{- type of survey MMB to use ('averaged' or 'raw')}

\item{M}{- assumed rate of natural mortality}

\item{t.sf}{- time from survey to fishery}

\item{t.fm}{- time from fishery to mating}

\item{hm.pot}{- handling mortality rate for fixed gear (pot) fisheries}

\item{hm.trl}{- handling mortality rate for trawl fisheries}

\item{pct.male}{- assumed male percentage}
}
\value{
list with MMB at mating and other time series
}
\description{
Function to calculate MMMB at mating.
}
\details{

}

