% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTheta.R
\name{calcTheta}
\alias{calcTheta}
\title{Calculate \eqn{\theta}, the average exploitation rate on MMB associated with discard mortality}
\usage{
calcTheta(mmbLst, assYr = 2014, n = 3)
}
\arguments{
\item{mmbLst}{\itemize{
\item output from \code{\link{calcMMBMating}}
}}

\item{assYr}{\itemize{
\item assesssment year
}}

\item{n}{\itemize{
\item number of years to include in average
}}
}
\value{
2-element vector with \eqn{\theta} calculated as
\enumerate{
\item \eqn{\theta = \frac{\sum_y M_{F_{y}}}{\sum_y MMB_{{fishing}_{y}}}}        \link{ratio of averages}
\item \eqn{\theta = \frac{1}{N}\sum_y \frac{M_{F_{y}}}{{MMB}_{{fishing}_{y}})}} \link{average of ratios}
}
where
\itemize{
\item \emph{N} = number of years over which to average
\item \eqn{M_{F_{y}}} = total discard-related mortality on mature males in year \emph{y}
\item \eqn{{MMB}_{{fishing}_{y}}} = MMB just prior to the (pulse) fishery
}
}
\description{
Function to calculate \eqn{\theta}, the average exploitation rate on MMB associated with discard mortality.
}
\details{
None.
}
