% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisheryYear.R
\name{fisheryYear}
\alias{fisheryYear}
\title{Returns the crab fishery year as a character string}
\usage{
fisheryYear(year)
}
\arguments{
\item{year}{- the year (as of July 1) of the fishery}
}
\value{
the fishery year as a character string. For 2014, this would be '2014/15'
}
\description{
Function to generate a character string representing a crab fishery year.
}
\details{
None.
}
