% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getPath2REM.R
\name{getPath2REM}
\alias{getPath2REM}
\title{Get the path to the ADMB random effects model (REM) used for survey averaging.}
\usage{
getPath2REM()
}
\value{
the path (as a string) to the ADMB model
}
\description{
Function to get the path to the ADMB random effects model (REM) used for survey averaging.
}
\details{


Opens a file selector for the user to select the ADMB model.
}

