% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRunCommands.R
\name{getRunCommands}
\alias{getRunCommands}
\title{Generate run commands for an ADMB model run}
\usage{
getRunCommands(
  os = "osx",
  path2model = getPath2REM(),
  pin = FALSE,
  hess = TRUE,
  mcmc = FALSE
)
}
\arguments{
\item{os}{\itemize{
\item 'win' or 'mac'
}}

\item{path2model}{\itemize{
\item 
}}

\item{pin}{\itemize{
\item 
}}

\item{hess}{\itemize{
\item 
}}

\item{mcmc}{\itemize{
\item 
}}
}
\value{
string with the run commands
}
\description{
Function to generate a script to make an ADMB model run
}
\details{
None.
}
