% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getSurveyData.R
\name{getSurveyData}
\alias{getSurveyData}
\title{Read the survey data from a csv file and return it as a dataframe.}
\usage{
getSurveyData(fn = NULL)
}
\arguments{
\item{fn}{- survey data filename}
}
\value{
dataframe with columns:
\itemize{
 \item year
 \item type (abundance/biomass)
 \item sex
 \item category
 \item units
 \item value
 \item cv
}
Input units are converted to 1000's t (biomass) and millions (abundance).
}
\description{
Function to read the survey data from a csv file and return it as a dataframe.
}
\details{

}

