% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/surveyAveraging.InvVar.R
\name{surveyAveraging.InvVar}
\alias{surveyAveraging.InvVar}
\title{Smooth survey data using inverse-variance averaging}
\usage{
surveyAveraging.InvVar(srvData, type = "biomass", sex = "male",
  category = "mature", n = 3, pdfType = "lognormal", ci = 0.95)
}
\arguments{
\item{srvData}{- survey data dataframe}
}
\value{
dataframe with smoothed survey data
}
\description{
Function to smooth survey data using inverse-variance averaging
}
\details{

}

