% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/surveyAveraging.REM.R
\name{surveyAveraging.REM}
\alias{surveyAveraging.REM}
\title{Smooth/interpolate survey data using a random effects(kalman filter) model (REM).}
\usage{
surveyAveraging.REM(srvData, type = "biomass", sex = "male",
  category = "mature", pdfType = "lognormal", ci = 0.95,
  modelPath = getPath2REM(), verbose = FALSE, showPlot = FALSE)
}
\arguments{
\item{srvData}{- raw survey data dataframe}

\item{type}{- data type ('abundance' or 'biomass') to average}

\item{sex}{- sex ('male' or 'female') to average}

\item{category}{- category ('immature','mature', or 'legal') to average}

\item{pdfType}{- distribution for CIs}

\item{ci}{- confidence interval for CIs}

\item{modelPath}{- path to ADMB REM used for survey averaging}

\item{verbose}{- flag (T/F) to print intermediate output}

\item{showPlot}{- flag (T/F) to plot results}
}
\value{
dataframe with smoothed survey data
}
\description{
Function to smooth/interpolate survey data using a random effects(kalman filter) model (REM).
}
\details{


This function uses an ADMB random effects model (originally developed by Jim Ianelli
and subsequently modified by William Stockhausen) to smooth/interpolate survey data.

smooth survey data
}

