% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBmsy.R
\name{calcBmsy}
\alias{calcBmsy}
\title{Calculate \eqn{B_{MSY}} by averaging over specified time interval}
\usage{
calcBmsy(mmb, years = c(1980:1984, 1990:1997))
}
\arguments{
\item{mmb}{- time series of mmb at mating to average}

\item{years}{- years to incude in average}
}
\value{
\eqn{B_{MSY}} (in t)
}
\description{
Function to calculate \eqn{B_{MSY}} by averaging over specified time interval.
}
\details{
None.
}
