% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCIs.R
\name{calcCIs}
\alias{calcCIs}
\title{Compute confidence intervals}
\usage{
calcCIs(vals, cvs, pdfType = "lognormal", ci = 0.95, verbose = FALSE)
}
\arguments{
\item{vals}{\itemize{
\item vector of values
}}

\item{cvs}{\itemize{
\item vector of cvs
}}

\item{pdfType}{\itemize{
\item probability distribution for error bars
}}

\item{ci}{\itemize{
\item confidence interval for error bar plots
}}

\item{verbose}{\itemize{
\item flag (T/F) to print intermediate output
}}
}
\value{
list with vectors lci, uci as  named elements
}
\description{
Function to compute confidence intervals.
}
\details{
None.
}
