% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcFofl.R
\name{calcFofl}
\alias{calcFofl}
\title{Calculate Fofl using the harvest control rule.}
\usage{
calcFofl(mmb, Bmsy, maxFofl, alpha, beta)
}
\arguments{
\item{mmb}{- MMB at mating}

\item{Bmsy}{- Bmsy}

\item{maxFofl}{- max allowed Fofl}

\item{alpha}{- x-intercept of sloping control line}

\item{beta}{- threshold mmb/Bmsy to allow a directed fishery}
}
\value{
the value of Fofl given the mmb/Bmsy ratio, using the control rule.
}
\description{
Functio to calculate Fofl using the harvest control rule.
}
\details{

}

