% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcMMBMating.R
\name{calcMMBMating}
\alias{calcMMBMating}
\title{Calculate projected MMMB at mating.}
\usage{
calcMMBMating(fshData, avgSrvData, avgType = "IV", M = 0.18, t.sf = 3/12,
  t.fm = 4/12, hm.pot = 0.5, hm.trl = 0.8, pct.male = 0.5,
  showPlot = FALSE)
}
\arguments{
\item{fshData}{- dataframe with fishery data}

\item{avgSrvData}{- dataframe with MMB at time of survey}

\item{avgType}{- type of survey MMB to use ('raw', 'IV' or 'REM')}

\item{M}{- assumed rate of natural mortality}

\item{t.sf}{- time from survey to fishery}

\item{t.fm}{- time from fishery to mating}

\item{hm.pot}{- handling mortality rate for fixed gear (pot) fisheries}

\item{hm.trl}{- handling mortality rate for trawl fisheries}

\item{pct.male}{- assumed male percentage}
}
\value{
list with MMB at mating and other time series. Elements are:
\itemize{
 \item mmbMat = MB at maturity
 \item mmbFsh = MMB at fishery time
 \item mmbSrv = MMB at survey time
 \item retM = retained mortality
 \item dscM = list with elements:
 \itemize{
     \item tot = total discard mortality
     \item gft = groundfish trawl fisheries mortality
     \item gfp = groundfish pot fisheries mortality
     \item crb = crab fisheries discard mortality
 }
 \item plots = list with elements:
 \itemize{
     \item MMB = MMB time series ggplot2 plot object
 }
}
}
\description{
Function to calculate MMMB at mating.
}
\details{

}

