% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calcOFL.R
\name{calcOFL}
\alias{calcOFL}
\title{Calculate the OFL.}
\usage{
calcOFL(mmbSrvCurr, Bmsy, theta, M = 0.18, gamma = 1, alpha = 0.1,
  beta = 0.25, t.sf = 3/12, t.fm = 4/12)
}
\description{
Function to calculate the OFL.
}
\details{

}

