% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcPrjMMB.R
\name{calcPrjMMB}
\alias{calcPrjMMB}
\title{Calculate projected MMMB at mating.}
\usage{
calcPrjMMB(mmbSrvCurr, Fm, theta, M = 0.18, t.sf = 3/12, t.fm = 4/12)
}
\arguments{
\item{mmbSrvCurr}{- current MMB at time of survey}

\item{Fm}{- assumed fishing mortality rate}

\item{theta}{- ratio of discard mortality to MMB at fishing}

\item{M}{- assumed rate of natural mortality}

\item{t.sf}{- time from survey to fishery}

\item{t.fm}{- time from fishery to mating}

\item{pct.male}{- assumed male percentage}
}
\value{
list with elements:
\itemize{
 \item mmb=mmbMatPrj
 \item retM=prjRet,
 \item dscM=prjDsc
}
}
\description{
Function to calculate projected MMMB at mating.
}
\details{

}

