% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calcTheta.R
\name{calcTheta}
\alias{calcTheta}
\title{Calculate theta, the average exploitation rate on MMB}
\usage{
calcTheta(mmbLst, assYr = 2014, n = 3)
}
\arguments{
\item{mmbLst}{- output from calcMMBMating}

\item{assYr}{- assesssment year}

\item{n}{- number of years to average}
}
\value{
2-element vector with theta calculated as
\enumerate{
 \item ratio of averages
 \item average of ratios
}
}
\description{
Function to calculate theta, the average exploitation rate on MMB
}
\details{

}

