% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFisheryData.R
\name{getFisheryData}
\alias{getFisheryData}
\title{Read the fishery data from a csv file and return it as a dataframe}
\usage{
getFisheryData(fn = NULL)
}
\arguments{
\item{fn}{\itemize{
\item fishery data filename
}}
}
\value{
dataframe with columns
\itemize{
\item year
\item type (abundance/biomass)
\item fishery
\item gear (pot/trawl)
\item units
\item category
\item catch
}
Input units are converted to t (biomass) and millions (abundance).
}
\description{
Function to read the fishery data from a csv file and return it as a dataframe.
}
\details{
None.
}
