% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getRunCommands.R
\name{getRunCommands}
\alias{getRunCommands}
\title{Generate run commands for an ADMB model run}
\usage{
getRunCommands(os = "osx", path2model = getPath2REM(), pin = FALSE,
  hess = TRUE, mcmc = FALSE)
}
\arguments{
\item{os}{- 'win' or 'mac'}

\item{path2model}{-}

\item{pin}{-}

\item{hess}{-}

\item{mcmc}{-}
}
\description{
Function to generate a script to make an ADMB model run
}
\details{

}

