% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getSurveyData.R
\name{getSurveyData}
\alias{getSurveyData}
\title{Read the survey data from a csv file and return it as a dataframe.}
\usage{
getSurveyData(fn = NULL)
}
\arguments{
\item{fn}{- survey data filename}
}
\value{
dataframe with columns
year, type (abundance/biomass), sex, category, units, value, cv.
Input units are converted to 1000's t (biomass) and millions (abundance).
}
\description{
Function to read the survey data from a csv file and return it as a dataframe.
}
\details{

}

