% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotAvgdData.R
\name{plotAvgdData}
\alias{plotAvgdData}
\title{Plot averaged data.}
\usage{
plotAvgdData(dfr, yr2 = 1990)
}
\arguments{
\item{dfr}{- dataframe output from one of the surveyAveraging functions}

\item{yr2}{- year to start 2nd plot with}
}
\value{
none
}
\description{
Function to plot averaged (and raw) data using output from one of the surveyAveraging functions.
}
\details{

}

