% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAvgdData.R
\name{plotAvgdData}
\alias{plotAvgdData}
\title{Plot averaged data.}
\usage{
plotAvgdData(dfr, yr2 = 1990, maxY = NULL, verbose = FALSE,
  showPlot = FALSE)
}
\arguments{
\item{dfr}{- dataframe output from one of the surveyAveraging functions}

\item{yr2}{- starting year for 2nd arithmetic-scale plot}

\item{maxY}{- optional max scale for arithmetic-scale y axes}

\item{verbose}{- flag (T/F) to print intermediate output}

\item{showPlot}{- flag (T/F) to show plots immediately}
}
\value{
list with ggplot2 objects as named elements 
\itemize{
 \item arScl1 - arithmetic scale plot
 \item arScl2 - arithmetic scale plot, sstarting at yr2
 \item lnScl = log-scale plot
}
}
\description{
Function to plot averaged (and raw) data using output from one of the surveyAveraging functions.
}
\details{
Biomass units are assumed to be in t, abundance units in ones.
}

