% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotRawData.R
\name{plotRawData}
\alias{plotRawData}
\title{Plot raw survey and fishery data.}
\usage{
plotRawData(srvData = NULL, fshData = NULL, pdfType = "lognormal",
  ci = 0.95)
}
\arguments{
\item{srvData}{- survey data dataframe, path to csv file, or NULL}

\item{fshData}{- fisheries data dataframe, path to csv file, or NULL}

\item{pdfType}{- probability distribution for error bars}

\item{ci}{- confidence interval for error bar plots}
}
\value{
list with srvData, fshData as elements. For survey data, abundance is in
milllions and biomass is in 1000's t. For fishery data, catch biomass is in t.
}
\description{
Function to plot raw survey and fishery data.
}
\details{

}

