% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotRawData.R
\name{plotRawData}
\alias{plotRawData}
\title{Plot raw survey and fishery data.}
\usage{
plotRawData(srvData = NULL, fshData = NULL, yr2 = 1990,
  pdfType = "lognormal", ci = 0.95, verbose = FALSE, showPlot = FALSE)
}
\arguments{
\item{srvData}{- survey data dataframe, path to csv file, or NULL}

\item{fshData}{- fisheries data dataframe, path to csv file, or NULL}

\item{yr2}{- starting year for 2nd arithmetic scale plot}

\item{pdfType}{- probability distribution for error bars}

\item{ci}{- confidence interval for error bar plots}

\item{verbose}{- flag (T/F) to print diagnostics, etc}

\item{showPlot}{- flag (T/F) to show plots immediately}
}
\value{
list with named elements
\itemize{
 \item srvData - dataframe w/ abundance is in ones and biomass is in t
 \item fshData - dataframe w/ catch biomass is in t
 \item plots - list with ggplot2 objects
 \itemize{
     \item pS1
     \item pS2
     \item pS3
     \item pF1
     \item pF2
 }
}
}
\description{
Function to plot raw survey and fishery data.
}
\details{


Biomass units are assumed to be in t, abundance units in ones.
}

