% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/surveyAveraging.InvVar.R
\name{surveyAveraging.InvVar}
\alias{surveyAveraging.InvVar}
\title{Smooth survey data using inverse-variance averaging}
\usage{
surveyAveraging.InvVar(srvData, type = "biomass", sex = "male",
  category = "mature", n = 3, pdfType = "lognormal", ci = 0.95,
  verbose = FALSE, showPlot = FALSE)
}
\arguments{
\item{srvData}{- raw survey data dataframe}

\item{type}{- data type ('abundance' or 'biomass') to average}

\item{sex}{- sex ('male' or 'female') to average}

\item{category}{- category ('immature','mature', or 'legal') to average}

\item{n}{- number of years to average across}

\item{pdfType}{- distribution for CIs}

\item{ci}{- confidence interval for CIs}

\item{verbose}{- flag (T/F) to print intermediate output}

\item{showPlot}{- flag (T/F) to plot results}
}
\value{
dataframe with smoothed survey data
}
\description{
Function to smooth survey data using inverse-variance averaging
}
\details{

}

